move validator to fix production build
